functions/requirements.txt
# Firebase Cloud Functions Requirements
firebase-admin==6.4.0
firebase-functions==0.4.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Utility Libraries
python-dateutil==2.8.2

# Vertex AI for pattern detection
google-cloud-aiplatform>=1.38.0
